Added the ability to pull tag from the user_reg variable in controller.py.
Still need to incorporate this functionality when actual nodes are being
used. Also need to look at making the user_reg variable a set. Once
complete I need to take a look at the player database and make sure that
it includes all necessary entries for what a player can be expected to
want.